requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-chess>=1.10.0
pydantic-settings>=2.2.0 